requirement.txt
streamlit
google-generativeai
pypdf2
fpdf2